move server and client run to cmd
